+++ src/PSDocs/Configuration/ColumnPadding.cs
@@ -1,0 +1,13 @@
+﻿namespace PSDocs.Configuration
+{
+    public enum ColumnPadding : byte
+    {
+        None,
+
+        Single,
+
+        MatchHeader,
+
+        Undefined
+    }
+}