+++ src/PSDocs/Models/Alignment.cs
@@ -1,0 +1,13 @@
+﻿namespace PSDocs.Models
+{
+    public enum Alignment
+    {
+        Center,
+
+        Left,
+
+        Right,
+
+        Undefined
+    }
+}