+++ src/PSDocs/Models/TableColumnHeader.cs
@@ -1,0 +1,16 @@
+﻿namespace PSDocs.Models
+{
+    public sealed class TableColumnHeader
+    {
+        public TableColumnHeader()
+        {
+            Alignment = Alignment.Undefined;
+        }
+
+        public Alignment Alignment { get; set; }
+
+        public string Label { get; set; }
+
+        public int Width { get; set; }
+    }
+}