+++ tests/PSDocs.Tests/PSDocs.Common.Tests.ps1
@@ -177,6 +177,8 @@ Describe 'Invoke-PSDocument' -Tag 'FromPath' {
             $Error[0].Exception.Message | Should -Match '^(The term ''New-PSDocsInvalidCommand'' is not recognized as the name of a cmdlet)';
             { Invoke-PSDocument -Path $here -OutputPath $outputPath -Name InvalidCommandWithSection -ErrorAction Stop } | Should -Throw -ExceptionType PSDocs.Execution.InvokeDocumentException;
             $Error[0].Exception.Message | Should -Match '^(The term ''New-PSDocsInvalidCommand'' is not recognized as the name of a cmdlet)';
+            { Invoke-PSDocument -Path $here -OutputPath $outputPath -Name WithWriteError -ErrorAction Stop } | Should -Throw -ExceptionType PSDocs.Execution.InvokeDocumentException;
+            $Error[0].Exception.Message | Should -Match 'Verify Write-Error is raised as an exception';
         }
     }
 
@@ -260,6 +262,42 @@ Describe 'New-PSDocumentOption' -Tag 'Option' {
         It 'from YAML' {
             $option = New-PSDocumentOption -Option (Join-Path -Path $here -ChildPath 'PSDocs.Tests.yml');
             $option.Markdown.SkipEmptySections | Should -Be $False;
+        }
+    }
+
+    Context 'Read Markdown.ColumnPadding' {
+
+        It 'from default' {
+            $option = New-PSDocumentOption;
+            $option.Markdown.ColumnPadding | Should -Be 'MatchHeader';
+        }
+
+        It 'from Hashtable' {
+            $option = New-PSDocumentOption -Option @{ 'Markdown.ColumnPadding' = 'Single' };
+            $option.Markdown.ColumnPadding | Should -Be 'Single';
+        }
+
+        It 'from YAML' {
+            $option = New-PSDocumentOption -Option (Join-Path -Path $here -ChildPath 'PSDocs.Tests.yml');
+            $option.Markdown.ColumnPadding | Should -Be 'Single';
+        }
+    }
+
+    Context 'Read Markdown.UseEdgePipes' {
+
+        It 'from default' {
+            $option = New-PSDocumentOption;
+            $option.Markdown.UseEdgePipes | Should -Be 'WhenRequired';
+        }
+
+        It 'from Hashtable' {
+            $option = New-PSDocumentOption -Option @{ 'Markdown.UseEdgePipes' = 'Always' };
+            $option.Markdown.UseEdgePipes | Should -Be 'Always';
+        }
+
+        It 'from YAML' {
+            $option = New-PSDocumentOption -Option (Join-Path -Path $here -ChildPath 'PSDocs.Tests.yml');
+            $option.Markdown.UseEdgePipes | Should -Be 'Always';
         }
     }