+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - **Important change**: Renamed `-When` parameter on Section block to `-If`. `-When` is still supported
 - Added support for locked down environments to ensure that documents are executed as constrained code when Device Guard is used
   - Use the `Execution.LanguageMode = 'ConstrainedLanguage'` option to force constrained language mode
+- **Important change**: Improved markdown formatting for tables [#31](https://github.com/BernieWhite/PSDocs/issues/31)
+  - Table columns are now padded by default to match header width. Set `Markdown.ColumnPadding` option to `None` to match format style from PSDocs <= 0.5.0
+  - Pipe characters on the start and end of a table row are not added by default. Set `Markdown.UseEdgePipes` option to `Always` to match format style from PSDocs <= 0.5.0
+  - Property expressions now support Label, Expression, Alignment and Width keys
 - **Experimental**: Publishing of keywords for syntax completion with editors
 
 ## v0.5.0